refactor: complete rearrange of folder structure and removed admin view, for later reimplementation
