+++ src/assets/Appointments.css
@@ -13,6 +13,8 @@
 
 .appointmentMenu__buttonContainer {
   width: 20%;
+  display: flex;
+  flex-direction: column;
   min-width: 200px;
 }
