+++ src/assets/MyButton.css
@@ -9,6 +9,8 @@
   font-weight: 500;
   border-radius: 5px;
   border: solid 2px rgba(0, 0, 0, 0.2);
+  text-decoration: none;
+  color: black;
 }
 
 .MyButton:hover {
