--- src/views/Admin/Admin.css
@@ -1,21 +1,0 @@
-.button__panel {
-  max-width: 25em;
-  min-width: 20em;
-  margin-top: 40px;
-}
-.especial__group {
-  display: flex;
-  margin-bottom: 10px;
-}
-.form__group__especial {
-  border: solid 2px rgb(0, 0, 0, 0.2);
-  padding: 5px;
-  border-radius: 5px;
-  background: rgb(246, 246, 246);
-}
-.especial__group div {
-  width: 3em;
-}
-.especial__body input {
-  width: 100%;
-}
